feat(packages): half-credit costs, human validity labels, buy-ahead pays upfront
- Credits go decimal (0.5 steps): Course.credit_cost, Package.credits,
  StudentPackage totals, Booking.credits_deducted, ManualCreditGrant.amount
  — a 1.5-credit lesson no longer 500s on save. DRF now emits decimals as
  JSON numbers (COERCE_DECIMAL_TO_STRING=False); credit/cost inputs get
  step=0.5.
- Package cards show validity as '3 mesi' / '1 anno' instead of '90d'
  (school/HQ cards and student storefront, ICU plurals, 5 languages).
- Buy-ahead reworked per Carlo: the student PAYS IN FULL at purchase
  (no Stripe trial) and the credit window opens when the current package
  expires, running one billing interval from there. Stripe keeps billing
  on the purchase-date cycle while renewals roll the shifted window
  forward one interval at a time — every payment lands before the window
  it covers, exactly like the first. Dialog copy now says 'you pay now'.
  Webhook tests rewritten for the new lifecycle (5 tests).

+++ backend/catalog/models.py
@@ -51,7 +51,7 @@ class Course(UUIDTimeStampedModel):
     duration_minutes = models.IntegerField(default=60)
     max_capacity = models.IntegerField(default=10)
     reserve_spots = models.IntegerField(default=0)
-    credit_cost = models.IntegerField(default=1)
+    credit_cost = models.DecimalField(max_digits=5, decimal_places=1, default=1)  # half-credit steps allowed
     color = models.CharField(max_length=20, default="#6B1F3A")
     vip_booking_hours_before = models.IntegerField(default=0)
     min_booking_notice_hours = models.IntegerField(default=2)
@@ -120,7 +120,7 @@ class Package(UUIDTimeStampedModel):
     name_es = models.CharField(max_length=255, blank=True)
     description_it = models.TextField(blank=True)
     description_en = models.TextField(blank=True)
-    credits = models.IntegerField(default=10)
+    credits = models.DecimalField(max_digits=6, decimal_places=1, default=10)  # half-credit steps allowed
     validity_days = models.IntegerField(default=90)
     price = models.DecimalField(max_digits=10, decimal_places=2, default=0)
     # Legacy single-value restriction ("all" or one lesson-type id). Superseded
